refactor(maturity): extract the combine held-pool sub-section (#467)
First of the combine-workflow sub-section splits (kept small and independently
verifiable per the money-flow risk). Move the "Ví chờ gộp" held-settlement picker
out of MaturityResolveBody into `HeldPoolSection` in maturityResolveFields.tsx —
choosing which parked settlements to fold into the re-deposit. The sheet keeps the
held state (heldSel + derived selectedHeld/heldReceivedTotal used by the payload
model) and passes the pool + selection getter/setter + labels down. The now-unused
GitMerge import drops out of the sheet. No behavior or markup change.

Parity: the MaturityResolveSheet component tests + the maturity-combine-holding
E2E (the hold → consume / unhold / double-count-guard flow) both green.

Verified: full unit suite 1261 passed; typecheck 0 errors; lint 0 errors; E2E
maturity-combine-holding (7) passed. MaturityResolveSheet 1219 → 1195.
